==> 3-combine.ipynb <==
import pandas as pd
import numpy as np
import functools as ft
from math import radians, cos, sin, asin, sqrt

### Paper data (given in the papers_processed folder)

df_papers = [pd.read_excel('papers_processed/papers_final.xlsx')]
df_papers.append(pd.read_excel('papers_processed/gabriel.xlsx'))
df_papers.append(pd.read_excel('papers_processed/karmanya_main_results_new.xlsx'))
df_papers.append(pd.read_excel('papers_processed/maik.xlsx'))
df_papers.append(pd.read_excel('papers_processed/Paper16_data.xlsx'))
df_papers.append(pd.read_excel('papers_processed/Paper18_data.xlsx'))
df_papers.append(pd.read_excel('papers_processed/Paper21_data.xlsx'))
df_papers.append(pd.read_excel('papers_processed/Paper29_data.xlsx'))
df_papers.append(pd.read_excel('papers_processed/Paper30_data.xlsx'))
df_papers.append(pd.read_excel('papers_processed/Paper33_data.xlsx'))

### Final grid data with yield (given in the combined data folder)

df_data = pd.read_csv('combined data/soilgrids_agri4cast_yield.csv')

### Filtering data based on distance to its cloasest yield location

df_data = df_data[df_data['nearest_distance_km'] < 40]

### Changing the range of soil ph to 0-14 (0-140 befors)

df_data['phh2o_0-5cm_mean'], df_data['phh2o_5-15cm_mean'],df_data['phh2o_15-30cm_mean'],df_data['phh2o_30-60cm_mean'],df_data['phh2o_60-100cm_mean'],df_data['phh2o_100-200cm_mean'] = df_data['phh2o_0-5cm_mean']/10, df_data['phh2o_5-15cm_mean']/10,df_data['phh2o_15-30cm_mean']/10,df_data['phh2o_30-60cm_mean']/10,df_data['phh2o_60-100cm_mean']/10,df_data['phh2o_100-200cm_mean']/10

### Code responsible for processing numerical selection values (pH etc.)

def process_numerical(env_data, row, selection_env, temp):
    selection = row['selection']
    if "-" in selection:
        bounds = selection.split('-')
        
        satisfying_rows = env_data[(env_data[selection_env] > float(bounds[0]))]
        satisfying_rows = satisfying_rows[(satisfying_rows[selection_env] < float(bounds[1]))]
        satisfying_rows['results_detailed'], satisfying_rows['results_actual'], satisfying_rows['farming_practice'] = row['results_detailed'], row['results_actual'], row['farming_practice']
        temp.append(satisfying_rows)
        
    elif ">" in selection:
        bound = selection[1:]
        satisfying_rows = env_data[(env_data[selection_env] > float(bound))]
        satisfying_rows['results_detailed'], satisfying_rows['results_actual'], satisfying_rows['farming_practice'] = row['results_detailed'], row['results_actual'], row['farming_practice']
        temp.append(satisfying_rows)
    elif "<" in selection:
        bound = selection[1:]
        satisfying_rows = env_data[(env_data[selection_env] < float(bound))]
        satisfying_rows['results_detailed'], satisfying_rows['results_actual'], satisfying_rows['farming_practice'] = row['results_detailed'], row['results_actual'], row['farming_practice']
        temp.append(satisfying_rows)
    return temp

### This is just a map for soilGrids, which maps soil type names to ids

soil_type_map = {
    'Acrisols': 0,
    'Albeluvisols': 1,
    'Alisols': 2,
    'Andosols': 3,
    'Arenosols': 4,
    'Calcisols': 5,
    'Cambisols': 6,
    'Chernozems': 7,
    'Cryosols': 8,
    'Durisols': 9,
    'Ferralsols': 10,
    'Fluvisols': 11,
    'Gleysols': 12,
    'Gypsisols': 13,
    'Histosols': 14,
    'Kastanozems': 15,
    'Leptosols': 16,
    'Lixisols': 17,
    'Luvisols': 18,
    'Nitisols': 19,
    'Phaeozems': 20,
    'Planosols': 21,
    'Plinthosols': 22,
    'Podzols': 23,
    'Regosols': 24,
    'Solonchaks': 25,
    'Solonetz': 26,
    'Stagnosols': 27,
    'Umbrisols': 28,
    'Vertisols': 29
}

### Code responsible for processing soil type selection, either by clay, sand and silt content or scientific name

def filter_by_soil_type(env_data, row, temp):
    soil_type = row['selection']
    
    # Convert g/kg to percentage
    sand_pct = env_data['sand_0-5cm_mean'] / 10
    silt_pct = env_data['silt_0-5cm_mean'] / 10
    clay_pct = env_data['clay_0-5cm_mean'] / 10
    ph = env_data['phh2o_0-5cm_mean']
    condition = "THIS IS A WALLUE"
    # Define conditions for each soil type
    if soil_type.lower() == 'sandy':
        
        condition = (sand_pct > 85) 
    
    elif soil_type.lower() == 'loamy':
        condition = (
            (sand_pct <= 52) & 
            (silt_pct >= 28) & (silt_pct <= 50) & 
            (clay_pct >= 7) & (clay_pct <= 27)
        )
    
    elif soil_type.lower() == 'clay':
        condition = (sand_pct <= 45) & (silt_pct <= 40) & (clay_pct >= 40)
    elif soil_type.lower() == 'sandy,acidic':
        condition = (sand_pct > 85) & (silt_pct + 1.5 * clay_pct < 15) & (ph <7)
    elif soil_type.lower() == 'lixisols':
        satisfying_rows = env_data[env_data['MostProbable'] == soil_type_map['Lixisols']]
        satisfying_rows['results_detailed'], satisfying_rows['results_actual'], satisfying_rows['farming_practice'] = row['results_detailed'], row['results_actual'], row['farming_practice']
        temp.append(satisfying_rows)
        return temp
    elif soil_type.lower() == 'arenosols':
        satisfying_rows = env_data[env_data['MostProbable'] == soil_type_map['Arenosols']]
        satisfying_rows['results_detailed'], satisfying_rows['results_actual'], satisfying_rows['farming_practice'] = row['results_detailed'], row['results_actual'], row['farming_practice']
        temp.append(satisfying_rows)
        return temp
    elif soil_type.lower() == 'acrisols':
        satisfying_rows = env_data[env_data['MostProbable'] == soil_type_map['Acrisols']]
        satisfying_rows['results_detailed'], satisfying_rows['results_actual'], satisfying_rows['farming_practice'] = row['results_detailed'], row['results_actual'], row['farming_practice']
        temp.append(satisfying_rows)
        return temp
    else:
        print(
            f"Invalid soil_type '{soil_type}'. "
            "Must be 'sandy', 'loamy', or 'clay'"
        )
        return temp
    satisfying_rows = env_data[condition]
    satisfying_rows['results_detailed'], satisfying_rows['results_actual'], satisfying_rows['farming_practice'] = row['results_detailed'], row['results_actual'], row['farming_practice']
    temp.append(satisfying_rows)
    return temp

### Code responsible for processing region selections, by creating lat-lon bounding boxes on the region

def filter_by_region(env_data, row, temp):
    region = row['selection']
    
    
    lon = env_data['lon'] 
    lat = env_data['lat'] 
    
    condition = "THIS IS A WALLUE"
    
    if region.lower() == 'spain':
        condition = (lon >= -9.30) & (lon <= 3.33) & (lat >= -36.00) & (lat <= 43.80)
    
    elif region.lower() == 'turkey':
        condition = (lon >= 25.00) & (lon <= 45.00) & (lat >= 35.8) & (lat <= 42.1)
    else:
        print(
            f"Invalid region '{region}'. "
            "Must be 'spain', or 'turkey'"
        )
        return temp
    satisfying_rows = env_data[condition]
    satisfying_rows['results_detailed'], satisfying_rows['results_actual'], satisfying_rows['farming_practice'] = row['results_detailed'], row['results_actual'], row['farming_practice']
    temp.append(satisfying_rows)
    return temp

### Map between selection names in our paper Excel data and how they are called in the grid data

process_dict = {
    'yearly precipitation': 'yearly_rain_mean',
    'latitude': 'lat',
    'soil ph': 'phh2o_0-5cm_mean',
    'yearly temperature': 'yearly_avg_mean_temp_mean'
}

### This code calls the previous 3 functions to do the actual selection of grids

def combine(env_data, paper_data):
    temp = []
    for idx, row in paper_data.iterrows():
        sel_type = row['selection_type'].lower()
        if sel_type in process_dict.keys():
            temp = process_numerical(env_data, row, process_dict[sel_type], temp)
        elif sel_type == "soil_type":
            temp = filter_by_soil_type(env_data, row, temp)
        elif sel_type == "region":
            temp = filter_by_region(env_data, row, temp)
    combined = pd.DataFrame(columns=['results_detailed','farming_practice','results_actual'])   
    if len(temp) > 0:
        combined = pd.concat(temp, ignore_index=True)   
    return combined
            
                

##### Simple renaming

df_papers[4]['selection_type'] = df_papers[4]['selection type']

### Calling the combine function for each paper and than concatenating the grids into a single one

combined = []
for idx, data in enumerate(df_papers):
    combined.append(combine(df_data, data))
combined_fin = pd.concat(combined, ignore_index=True)  
combined_fin

### Saving the raw data after combining

combined_fin.to_csv('data_raw.csv', index=False)

### Processing one of the collumns to revieve the yield effect as a float, this will be used later to select the best farming practice per grid

test = pd.DataFrame()
test['results_actual'] = combined_fin["results_actual"].str.split(",")
for idx, row in test.iterrows():
    test.iloc[idx]['results_actual'] = test.iloc[idx]['results_actual'][0].strip('"\'')
test['results_actual'] = test['results_actual'].astype(float)
combined_fin['yield_effect'] = test['results_actual']

### Filtering for such that only the best farming practice is left per grid

data_with_yield_dist_only_best = combined_fin.loc[combined_fin.groupby(['lon', 'lat'])['yield_effect'].idxmax()]
data_with_yield_dist_only_best

### Saving the data that has been filtered for the best farming practice per grid

data_with_yield_dist_only_best.to_csv('data_distanceFilter_best.csv', index=False)
